SpotifyBackend/requirements.txt
Flask==2.1.1
spotipy==2.20.0
python-dotenv==0.19.2
